Scroll to top on logo click on home page (#341)
* Hompe page bottom redesign using table

* Made footer smaller

* A little refactoring and scroll to top on logo click on home page

* Fix padding blocking a part of logo
