+++ simplq/src/components/common/ClickableLogo.jsx
@@ -4,8 +4,14 @@ import React from 'react';
 import styles from '../../styles/logo.module.scss';
 
 export default (props) => {
+  let onClick = {};
+  if (props.history) {
+    onClick = () => props.history.push('/');
+  } else if (props.onClick) {
+    onClick = props.onClick;
+  }
   return (
-    <div className={styles['logo']} onClick={() => (props.history ? props.history.push('/') : {})}>
+    <div className={styles['logo']} onClick={onClick}>
       <img src="/LogoLight.png" alt="Home" />
       <p>SimplQ</p>
     </div>
