+++ simplq/src/components/pages/Home/Nav/LeftNav.jsx
@@ -2,32 +2,18 @@
 
 import React from 'react';
 import styles from '../../../../styles/homePage.module.scss';
+import { smoothScrollTo } from '../../../common/utilFns';
 
-function smoothScrollTo(targetElement) {
-  // offset to avoid the sticky header on top to block the "Create a Queue at the click of a button" text
-  const offset = 45;
-  // setting the exact position on the document to scroll to
-  const bodyRect = document.body.getBoundingClientRect().top;
-  const elementRect = targetElement.getBoundingClientRect().top;
-  const elementPosition = elementRect - bodyRect;
-  const offsetPosition = elementPosition - offset;
-  // scroll to the exact position
-  window.scrollTo({
-    top: offsetPosition,
-    behavior: 'smooth',
-  });
-}
-function scrollToHowItWorks() {
+const scrollToHowItWorks = () => {
   // get the target div by ID
   const element = document.getElementById('target_how_it_works');
   smoothScrollTo(element);
-}
+};
 
-function scrollToContactUs() {
-  // get the target div by ID
+const scrollToContactUs = () => {
   const element = document.getElementById('target_contact_us');
   smoothScrollTo(element);
-}
+};
 
 const LeftNav = ({ open }) => {
   return (
