+++ simplq/src/styles/homePage.module.scss
@@ -262,7 +262,7 @@
   a {
     text-decoration: none;
     color: grey;
-    cursor: auto;
+    cursor: pointer;
   }
   @media (max-width: 768px) {
     flex-flow: column nowrap;
