+++ simplq/src/styles/logo.module.scss
@@ -3,11 +3,10 @@
 .logo {
   display: flex;
   cursor: pointer;
-
+  z-index: 10;
   p {
     font-family: $header-font;
     color: $primary-color-dark;
-    cursor: pointer;
   }
   & > img {
     height: 3rem;
